+++ Assets/Scenes/MenuScene.unity
@@ -1015,14 +1015,35 @@ PrefabInstance:
       propertyPath: m_AnchoredPosition.y
       value: -35
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 5009541612151403793, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects:
     - targetCorrespondingSourceObject: {fileID: 7384440929180223686, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
         type: 3}
       insertIndex: -1
       addedObject: {fileID: 647212798}
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 5009541612151404013, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1198431899}
+    - targetCorrespondingSourceObject: {fileID: 5009541612151404013, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1198431898}
+    - targetCorrespondingSourceObject: {fileID: 5009541612151404013, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1198431897}
+    - targetCorrespondingSourceObject: {fileID: 5009541612151404013, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1198431896}
+    - targetCorrespondingSourceObject: {fileID: 5009541612151404013, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1198431900}
   m_SourcePrefab: {fileID: 100100000, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8, type: 3}
 --- !u!224 &874813133 stripped
 RectTransform:
@@ -1400,6 +1421,101 @@ RectTransform:
     type: 3}
   m_PrefabInstance: {fileID: 1142642072}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &1198431894 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 5009541612151404013, guid: 92c5908dee7ffd04ea10a8d5d5a15ab8,
+    type: 3}
+  m_PrefabInstance: {fileID: 837332342}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1198431896
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1198431894}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d8ace06a29e24742a57e7063a7345cf9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: Assembly-CSharp::YARG.Menu.MenuObject
+  <Menu>k__BackingField: 8
+  <HideBelow>k__BackingField: 1
+--- !u!114 &1198431897
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1198431894}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dc42784cf147c0c48a680349fa168899, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_IgnoreReversedGraphics: 1
+  m_BlockingObjects: 0
+  m_BlockingMask:
+    serializedVersion: 2
+    m_Bits: 4294967295
+--- !u!114 &1198431898
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1198431894}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0cd44c1031e13a943bb63640046fad76, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_UiScaleMode: 1
+  m_ReferencePixelsPerUnit: 100
+  m_ScaleFactor: 1
+  m_ReferenceResolution: {x: 1920, y: 1080}
+  m_ScreenMatchMode: 1
+  m_MatchWidthOrHeight: 0
+  m_PhysicalUnit: 3
+  m_FallbackScreenDPI: 96
+  m_DefaultSpriteDPI: 96
+  m_DynamicPixelsPerUnit: 1
+  m_PresetInfoIsWorld: 0
+--- !u!223 &1198431899
+Canvas:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1198431894}
+  m_Enabled: 1
+  serializedVersion: 3
+  m_RenderMode: 0
+  m_Camera: {fileID: 0}
+  m_PlaneDistance: 100
+  m_PixelPerfect: 0
+  m_ReceivesEvents: 1
+  m_OverrideSorting: 0
+  m_OverridePixelPerfect: 0
+  m_SortingBucketNormalizedSize: 0
+  m_VertexColorAlwaysGammaSpace: 0
+  m_AdditionalShaderChannelsFlag: 25
+  m_UpdateRectTransformForStandalone: 0
+  m_SortingLayerID: 0
+  m_SortingOrder: 0
+  m_TargetDisplay: 0
+--- !u!114 &1198431900
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1198431894}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6c772d469c695494c952c88a3d715ecc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: Assembly-CSharp::YARG.Menu.Archipelago.ArchipelagoMenu
 --- !u!1001 &1271105421
 PrefabInstance:
   m_ObjectHideFlags: 0
